refactor: moved ApproveButton and SelectAllButton components to a separate file
